requirements.txt
fastapi
uvicorn
opencv-python
ffmpeg-python
ultralytics
python-multipart
pydantic
ipykernel